Added directory parsing and error logging
Running the program with -d <dirpath> with recursively go through all files and
directories, converting all the tpl files to phtml.

Rather than dying on exception (which is if the syntax doesn't
exist in the grammer or if its wrong), it writes to a `convert_log` file with
the filename and error that occured.  Still needs some work on formatting and
consistancy but it's a start.